Add ProtoDescriptorInferrer and Test
This commit adds a `ProtoDescriptorInferrer` class to the `twister-proto`
module. This class is responsible for inferring a protobuf `Descriptor`
from a given `Map<String, Object>`.

The `ProtoDescriptorInferrerTest` is also added to verify the proper
functioning of `ProtoDescriptorInferrer`. The test uses a `Map<String, Object>`
with various field types and ensures that the inferred descriptor matches
the expected descriptor.

Other changes include minor clean-ups in the `AvroReaderTest` and
`ProtoReaderTest` classes in the `twister-avro` and `twister-proto` modules
respectively, where unused imports are removed.