Add Documentation to Avro and Proto Classes
In this commit, we've added JavaDoc comments to several classes in the
`twister-avro` and `twister-proto` packages. This provides a clearer
understanding of the purpose and use of each class and method.

In the `twister-avro` package, the `AvroReader`, `AvroSchemaInferrer`, and
`AvroWriter` classes were updated.

In the `twister-proto` package, we enhanced `ProtoDescriptorInferrer`,
`ProtoReader`, and `ProtoWriter` with detailed comments.

This commit will greatly improve code readability and maintainability by
providing necessary context to developers working with these classes.