Add `write` method overload in ProtoWriter and corresponding test
This commit adds a new overload for the `write` method in the `ProtoWriter` class,
which now accepts a message name string and infers the descriptor using a new
instance of `ProtoDescriptorInferrer`. A test case `testWriteWithInferredDescriptor`
has been added to the `ProtoWriterTest` class to ensure this new functionality is
correctly working.

Changes include:
* New `write` method overload in `ProtoWriter` with Javadoc documentation.
* New test case `testWriteWithInferredDescriptor` in `ProtoWriterTest`.
* Minor import optimization in `ProtoWriterTest`.