Add 'Getting Started' Section in README with Maven and Gradle Instructions
This commit adds a 'Getting Started' section to the README file. The new
section provides instructions on how to add `twister-avro` and `twister-proto`
dependencies to a project using either Maven or Gradle. The provided XML and
Groovy code snippets are clear and easy to follow, helping new users to quickly
get started with Twister.

This addition to the documentation will be beneficial to users who are new to
the Twister project, providing clear instructions on how to add Twister as a
dependency in their own projects.