Add New Parquet Features to Future Work in README
This commit expands the 'Future Work' section of the README file with two new
items related to the Parquet format. The features 'Parquet ➡️ Map POJOs' and
'Parquet ➡️ Map wrapper' have been added to the list.

The addition of these features aims to provide a clear roadmap for contributors
and users. It communicates our intent to extend the functionality of the
Twister library by offering support for the Parquet format.

In future updates, the Twister project aims to support Parquet to Map POJOs
and a Parquet to Map wrapper, thereby increasing the data formats that the
library can handle effectively.

a